+++ src/main/java/com/notenoughmail/kubejs_tfc/addons/entityjs/EventHandlers.java
@@ -1,7 +1,7 @@
 package com.notenoughmail.kubejs_tfc.addons.entityjs;
 
 import com.notenoughmail.kubejs_tfc.KubeJSTFC;
-import com.notenoughmail.kubejs_tfc.addons.entityjs.builders.IHaveFaunaDefinition;
+import com.notenoughmail.kubejs_tfc.addons.entityjs.builders.IFaunaDefinable;
 import net.dries007.tfc.common.entities.EntityHelpers;
 import net.dries007.tfc.common.entities.Fauna;
 import net.dries007.tfc.common.entities.aquatic.AquaticMob;
@@ -12,6 +12,7 @@ import net.minecraft.tags.BlockTags;
 import net.minecraft.world.level.chunk.ChunkGenerator;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.entity.SpawnPlacementRegisterEvent;
+import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.javafmlmod.FMLModContainer;
@@ -19,14 +20,13 @@ import net.minecraftforge.fml.javafmlmod.FMLModContainer;
 public class EventHandlers {
 
     public static void init() {
-        final IEventBus bus = MinecraftForge.EVENT_BUS;
         final IEventBus modBus = ((FMLModContainer) ModList.get().getModContainerById(KubeJSTFC.MODID).get()).getEventBus(); // I hate mod buses
 
-        modBus.addListener(EventHandlers::spawnPlacement);
+        modBus.addListener(EventPriority.LOWEST, EventHandlers::spawnPlacement);
     }
 
     private static void spawnPlacement(SpawnPlacementRegisterEvent event) {
-        for (IHaveFaunaDefinition.FaunaType<?> type : IHaveFaunaDefinition.registeredFaunas) {
+        for (IFaunaDefinable.FaunaType<?> type : IFaunaDefinable.registeredFaunas) {
             event.register(type.type().get(), type.place(), type.heightMap(), (mob, level, heightmap, pos, rand) -> {
                 final Fauna fauna = type.fauna().get();
                 final ChunkGenerator generator = level.getLevel().getChunkSource().getGenerator();
@@ -58,7 +58,7 @@ public class EventHandlers {
                     return false;
                 }
                 return fauna.getMaxBrightness() == -1 || level.getRawBrightness(pos, 0) <= fauna.getMaxBrightness();
-            }, SpawnPlacementRegisterEvent.Operation.AND);
+            }, SpawnPlacementRegisterEvent.Operation.REPLACE);
         }
     }
 }