+++ src/main/java/com/notenoughmail/kubejs_tfc/addons/entityjs/builders/IFaunaDefinable.java
@@ -1,6 +1,8 @@
 package com.notenoughmail.kubejs_tfc.addons.entityjs.builders;
 
+import dev.latvian.mods.kubejs.typings.Info;
 import dev.latvian.mods.rhino.util.HideFromJS;
+import dev.latvian.mods.rhino.util.RemapForJS;
 import net.dries007.tfc.common.entities.Fauna;
 import net.liopyu.entityjs.builders.living.BaseLivingEntityBuilder;
 import net.liopyu.entityjs.entities.living.entityjs.IAnimatableJS;
@@ -13,21 +15,22 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
 
-public interface IHaveFaunaDefinition<T extends LivingEntity & IAnimatableJS> extends Supplier<EntityType<T>> {
+public interface IFaunaDefinable {
 
     List<FaunaType<?>> registeredFaunas = new ArrayList<>();
 
-    default BaseLivingEntityBuilder<T> withFaunaDefinition(SpawnPlacements.Type placementType, Heightmap.Types heightMap) {
-        registeredFaunas.add(new FaunaType<>(this, Fauna.MANAGER.register(self().id), placementType, heightMap));
-        return self();
+    @Info(value = "Registers a TFC fauna definition and overrides the entity's spawn placement to use it")
+    @RemapForJS("withFaunaDefinition")
+    default BaseLivingEntityBuilder<?> kubejs_tfc$WithFaunaDefinition(SpawnPlacements.Type placementType, Heightmap.Types heightMap) {
+        registeredFaunas.add(new FaunaType<>(kubejs_tfc$Self(), Fauna.MANAGER.register(kubejs_tfc$Self().id), placementType, heightMap));
+        return kubejs_tfc$Self();
     }
 
     @HideFromJS
-    default BaseLivingEntityBuilder<T> self() {
+    default <T extends LivingEntity & IAnimatableJS> BaseLivingEntityBuilder<T> kubejs_tfc$Self() {
         return (BaseLivingEntityBuilder<T>) this;
     }
 
     // TFC's version of this is not accessible, so just duplicate it
-    // TODO: [1.2.0] Possible event for registering a fauna for arbitrary entities?
     record FaunaType<T extends LivingEntity>(Supplier<EntityType<T>> type, Supplier<Fauna> fauna, SpawnPlacements.Type place, Heightmap.Types heightMap) {}
 }