+++ src/main/java/com/notenoughmail/kubejs_tfc/addons/entityjs/entities/OviparousAnimalJS.java
@@ -1,10 +1,10 @@
 package com.notenoughmail.kubejs_tfc.addons.entityjs.entities;
 
-import com.notenoughmail.kubejs_tfc.addons.entityjs.builders.MammalJSBuilder;
-import net.dries007.tfc.common.entities.livestock.Mammal;
-import net.dries007.tfc.common.entities.livestock.TFCAnimal;
+import com.notenoughmail.kubejs_tfc.addons.entityjs.builders.OviparousAnimalJSBuilder;
+import net.dries007.tfc.common.entities.livestock.OviparousAnimal;
 import net.liopyu.entityjs.builders.living.BaseLivingEntityBuilder;
 import net.liopyu.entityjs.entities.living.entityjs.IAnimatableJS;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Item;
@@ -12,13 +12,13 @@ import net.minecraft.world.level.Level;
 import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
-public class MammalJS extends Mammal implements IAnimatableJS {
+public class OviparousAnimalJS extends OviparousAnimal implements IAnimatableJS {
 
+    private final OviparousAnimalJSBuilder builder;
     private final AnimatableInstanceCache animationCache;
-    private final MammalJSBuilder builder;
 
-    public MammalJS(EntityType<? extends TFCAnimal> animal, Level level, MammalJSBuilder builder) {
-        super(animal, level, builder.sounds, builder.config.mammal());
+    public OviparousAnimalJS(EntityType<? extends OviparousAnimal> type, Level level, OviparousAnimalJSBuilder builder) {
+        super(type, level, builder.sounds, builder.config.oviparous(), builder.crows);
         this.builder = builder;
         animationCache = GeckoLibUtil.createInstanceCache(this);
     }
@@ -36,5 +36,10 @@ public class MammalJS extends Mammal implements IAnimatableJS {
     @Override
     public AnimatableInstanceCache getAnimatableInstanceCache() {
         return animationCache;
+    }
+
+    @Override
+    public MutableComponent getProductReadyName() {
+        return builder.productReadyMessage == null ? super.getProductReadyName() : builder.productReadyMessage;
     }
 }