+++ src/main/java/com/notenoughmail/kubejs_tfc/addons/entityjs/entities/WoolyAnimalJS.java
@@ -1,10 +1,10 @@
 package com.notenoughmail.kubejs_tfc.addons.entityjs.entities;
 
-import com.notenoughmail.kubejs_tfc.addons.entityjs.builders.MammalJSBuilder;
-import net.dries007.tfc.common.entities.livestock.Mammal;
-import net.dries007.tfc.common.entities.livestock.TFCAnimal;
+import com.notenoughmail.kubejs_tfc.addons.entityjs.builders.WoolAnimalJSBuilder;
+import net.dries007.tfc.common.entities.livestock.WoolyAnimal;
 import net.liopyu.entityjs.builders.living.BaseLivingEntityBuilder;
 import net.liopyu.entityjs.entities.living.entityjs.IAnimatableJS;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Item;
@@ -12,13 +12,13 @@ import net.minecraft.world.level.Level;
 import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
-public class MammalJS extends Mammal implements IAnimatableJS {
+public class WoolyAnimalJS extends WoolyAnimal implements IAnimatableJS {
 
+    private final WoolAnimalJSBuilder builder;
     private final AnimatableInstanceCache animationCache;
-    private final MammalJSBuilder builder;
 
-    public MammalJS(EntityType<? extends TFCAnimal> animal, Level level, MammalJSBuilder builder) {
-        super(animal, level, builder.sounds, builder.config.mammal());
+    public WoolyAnimalJS(EntityType<? extends WoolyAnimal> animal, Level level, WoolAnimalJSBuilder builder) {
+        super(animal, level, builder.sounds, builder.config.producingMammal());
         this.builder = builder;
         animationCache = GeckoLibUtil.createInstanceCache(this);
     }
@@ -26,6 +26,11 @@ public class MammalJS extends Mammal implements IAnimatableJS {
     @Override
     public TagKey<Item> getFoodTag() {
         return builder.food;
+    }
+
+    @Override
+    public MutableComponent getProductReadyName() {
+        return builder.productMessage == null ? super.getProductReadyName() : builder.productMessage;
     }
 
     @Override